refactor(T1 inc2): single date-generic shared domain-types module
Extract the core sign-in domain shapes (Station, Member, Activity, CheckIn,
ActiveActivity, Event, EventParticipant, EventWithParticipants, plus
CheckInMethod/StationHierarchy) into one date-generic, declaration-only module
shared/domain-types.d.ts (Member<TDate = string> etc.).

- backend/src/types/index.ts re-exports them specialised with Date
- frontend/src/types/index.ts re-exports them with the default string
- App-specific composites (StationLookupResult, truck-check types) and the
  frontend-only CheckInWithDetails.tagColor / ActiveActivity.activity extensions
  stay app-local

Mechanism deliberately avoids the npm-workspace foundation (T6) that conflicts
with the per-app deploy: a type-only .d.ts is never emitted, so backend/dist and
frontend/dist are structurally unchanged and CI deploy packaging is untouched
(no package.json / node_modules / workspace changes).

Verified: backend build emits dist/index.js unchanged with no shared/ leak;
all CI gates green (680 backend + 453 frontend + 85 AAR tests; lint + typechecks).
